Remove the superfluous Graphic class
The Graphic class was created to extend it in other classes, but after creating the IsometricGraphic class, the previous one doesn't have any function, it has been removed and all its declarations have been declared in the IsometricGraphic class
