+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.4.1] - 2020-09-26
+
+- Remove the superfluous Graphic class
+
 ## [2.4.0] - 2020-05-25
 
 - Created a new IsometricCircle class to build isometric circles
