--- src/@classes/abstract/Graphic/index.ts
@@ -1,2 +1,0 @@
-export * from './Graphic';
-export * from './types';
