+++ src/@classes/abstract/IsometricGraphic/index.ts
@@ -1,1 +1,2 @@
 export { IsometricGraphic } from './IsometricGraphic';
+export { IsometricGraphicProps } from './types';
