+++ src/index.ts
@@ -1,6 +1,6 @@
 export { IsometricCanvas, IsometricCanvasProps } from '@classes/public/IsometricCanvas';
 export { IsometricPath } from '@classes/public/IsometricPath';
-export { GraphicProps as IsometricPathProps } from '@classes/abstract/Graphic';
+export { IsometricGraphicProps } from '@classes/abstract/IsometricGraphic';
 export { IsometricRectangle, IsometricRectangleProps } from '@classes/public/IsometricRectangle';
 export { IsometricCircle, IsometricCircleProps } from '@classes/public/IsometricCircle';
 export { LineCap, LineJoin, PlaneView } from '@types';
